@@ -1,8 +1,10 @@
-﻿using Orleans.Services;
+﻿using System.Threading.Tasks;
+using Orleans.Services;
 
-namespace net.jommy.RuuviCore.Interfaces
+namespace net.jommy.RuuviCore.Interfaces;
+
+public interface IRuuviDBusListener : IGrainService
 {
-    public interface IRuuviDBusListener : IGrainService
-    {
-    }
+    [Orleans.Alias("SimulateEvent")]
+    Task SimulateEvent(string macAddress);
 }
