@@ -1,11 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
+
 using InfluxDB.Collector;
 using InfluxDB.Collector.Diagnostics;
+
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+
 using net.jommy.RuuviCore.Common;
 using net.jommy.RuuviCore.Interfaces;
+
 using Orleans;
 using Orleans.Concurrency;
 
@@ -15,48 +21,75 @@ namespace net.jommy.RuuviCore.Grains;
 public class InfluxBridge : Grain, IInfluxBridge
 {
     private MetricsCollector _metricsCollector;
-    private InfluxSettings _influxSettings;
-    private readonly IInfluxSettingsFactory _influxSettingsFactory;
+    private readonly InfluxSettings _influxSettings;
     private readonly ILogger<InfluxBridge> _logger;
 
-    public InfluxBridge(IInfluxSettingsFactory influxSettingsFactory, ILogger<InfluxBridge> logger)
+    public InfluxBridge(IOptions<InfluxSettings> influxOptions, ILogger<InfluxBridge> logger)
     {
-        _influxSettingsFactory = influxSettingsFactory;
+        _influxSettings = influxOptions.Value;
         _logger = logger;
     }
 
-    public override Task OnActivateAsync()
+    public override Task OnActivateAsync(CancellationToken cancellationToken)
     {
-        _influxSettings = _influxSettingsFactory.GetSettings(this.GetPrimaryKeyString());
+        _logger.LogDebug(
+            "Activating InfluxDB bridge to {InfluxAddress}, database: {InfluxDatabase}, measurement table: {InfluxMeasurementTable}, UserName: {UserName}, Password: {Password}.",
+            _influxSettings.InfluxAddress,
+            _influxSettings.InfluxDatabase,
+            _influxSettings.InfluxMeasurementTable,
+            _influxSettings.Username,
+            MaskPassword(_influxSettings.Password));
         _metricsCollector = new CollectorConfiguration()
             .Batch.AtInterval(TimeSpan.FromSeconds(2))
-            .WriteTo.InfluxDB(_influxSettings.InfluxAddress, _influxSettings.InfluxDatabase, _influxSettings.Username, _influxSettings.Password)
+            .WriteTo.InfluxDB(
+                _influxSettings.InfluxAddress,
+                _influxSettings.InfluxDatabase,
+                _influxSettings.Username,
+                _influxSettings.Password)
             .CreateCollector();
         CollectorLog.RegisterErrorHandler((message, exception) =>
         {
-            _logger.LogError(exception, "Error when activating bridge to influx: {errorMessage}", message);
+            _logger.LogError(
+                "{Message}. Error when activating bridge to influx: {ErrorMessage}",
+                message,
+                exception?.Message);
         });
-            
-        return base.OnActivateAsync();
+
+        return Task.CompletedTask;
     }
 
-    public Task<bool> WriteMeasurements(string macAddress, string name, Measurements measurements)
+    private static string MaskPassword(string influxSettingsPassword)
+    {
+        if (string.IsNullOrEmpty(influxSettingsPassword))
+        {
+            return "<empty>";
+        }
+
+        return influxSettingsPassword[0] + new string('*', influxSettingsPassword.Length - 1);
+    }
+
+    public Task<bool> WriteMeasurements(string macAddress, string name, MeasurementDTO measurements)
     {
         try
         {
-            var tags = new Dictionary<string, string> { {"mac", macAddress}};
+            var tags = new Dictionary<string, string> { { "mac", macAddress } };
             if (name != null)
             {
                 tags["name"] = name;
             }
-            _metricsCollector.Write(_influxSettings.InfluxMeasurementTable, new Dictionary<string, object>
-            {
-                {"Temperature", measurements.Temperature},
-                {"Humidity", measurements.Humidity},
-                {"Pressure", measurements.Pressure},
-                {"Rssi", measurements.RSSI},
-                {"BatteryVoltage", measurements.BatteryVoltage}
-            }, tags, measurements.Timestamp);
+
+            _metricsCollector.Write(
+                _influxSettings.InfluxMeasurementTable,
+                new Dictionary<string, object>
+                {
+                    { "Temperature", measurements.Temperature },
+                    { "Humidity", measurements.Humidity },
+                    { "Pressure", measurements.Pressure },
+                    { "Rssi", measurements.RSSI },
+                    { "BatteryVoltage", measurements.BatteryVoltage }
+                },
+                tags,
+                measurements.Timestamp);
         }
         catch (Exception e)
         {
@@ -65,10 +98,5 @@ public class InfluxBridge : Grain, IInfluxBridge
         }
 
         return Task.FromResult(true);
-    }
-
-    public Task<bool> IsValid()
-    {
-        return Task.FromResult(_influxSettings.BridgeName == this.GetPrimaryKeyString());
     }
 }
